Fully switch over to the new Stripe integration
