+++ integrations/integration-stripe/index.ts
@@ -1,5 +1,4 @@
 // codegen:start {preset: barrel, include: "./{*.{ts,tsx},*/index.{ts,tsx}}", exclude: "./**/*.{d,spec,test,fixture,gen,node}.{ts,tsx}"}
-export * from './stripe-utils'
 export * from './StripeClient'
-export * from './StripeProvider'
+export * from './StripeIntegration'
 // codegen:end
