+++ integrations/integration-stripe/package.json
@@ -13,8 +13,7 @@
     "@usevenice/cdk-core": "workspace:*",
     "@usevenice/cdk-ledger": "workspace:*",
     "@usevenice/util": "workspace:*",
-    "react": "*",
-    "stripe": "12.2.0"
+    "react": "*"
   },
   "devDependencies": {
     "@types/react": "*",
